api/requirements.txt
flask
flask-cors
flask-limiter
google-genai
python-dotenv
werkzeug
